third_party/blink/web_tests/fast/events/wheel/continuous-platform-wheelevent-in-scrolling-div.html

<!DOCTYPE html>
<script src='../../../resources/gesture-util.js'></script>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../virtual/percent-based-scrolling/resources/percent-based-util.js"></script>
<div id="overflow" style="border:2px solid black;overflow:auto;height:200px;width:200px;">
    <div style="background-color:red;height:200px;width:400px;"></div>
    <div style="background-color:green;height:200px;width:400px;"></div>
    <div style="background-color:red;height:200px;width:400px;"></div>
</div>

<script>

const overflowElement = document.getElementById("overflow");
let last_event = null;
overflowElement.addEventListener("mousewheel", mousewheelHandler, false);

function mousewheelHandler(e) {
    last_event = e;
}
const SOURCE = GestureSourceType.MOUSE_INPUT;

const SCROLL_LEFT = 100;
const SCROLL_TOP = 200;
const PIXELS_TO_SCROLL = calculatePixelsToScroll(overflowElement, SCROLL_LEFT, SCROLL_TOP);
const EXPECTED_WHEEL_DELTA_X = PIXELS_TO_SCROLL.x / pixelsPerTick() *
                               LEGACY_MOUSE_WHEEL_TICK_MULTIPLIER * window.devicePixelRatio;
const EXPECTED_WHEEL_DELTA_Y = PIXELS_TO_SCROLL.y / pixelsPerTick() *
                               LEGACY_MOUSE_WHEEL_TICK_MULTIPLIER * window.devicePixelRatio;

promise_test(async () => {
    await percentBasedSmoothScroll(SCROLL_LEFT, 100, 110, SOURCE,
                        'right', SPEED_INSTANT, overflowElement);
    await waitForAnimationEndTimeBased(() => {
      return overflowElement.scrollLeft;
    });
    assert_approx_equals(overflowElement.scrollLeft, SCROLL_LEFT, 0.2);
    assert_equals(last_event.wheelDeltaX, -Math.floor(EXPECTED_WHEEL_DELTA_X));
    assert_equals(last_event.wheelDelta, -Math.floor(EXPECTED_WHEEL_DELTA_X));

    last_event = null;

    await percentBasedSmoothScroll(SCROLL_TOP, 100, 110, SOURCE, 'down',
                        SPEED_INSTANT, overflowElement);
    await waitForAnimationEndTimeBased(() => {
      return overflowElement.scrollTop;
    });
    assert_approx_equals(overflowElement.scrollTop, SCROLL_TOP, 0.2);
    assert_equals(last_event.wheelDeltaY, -Math.floor(EXPECTED_WHEEL_DELTA_Y));
    assert_equals(last_event.wheelDelta, -Math.floor(EXPECTED_WHEEL_DELTA_Y));
}, 'This test checks the wheel delta value of wheel events, which should be the number of ticks multiplies the legacy mouse wheel tick multiplier.');
</script>
